Move schema package from stargate.db.datastore to stargate.db
This is an automatic refactor. The rational is that the schema will
primarily move to `Persistence`, so won't be specific to datastore.
