+++ graphqlapi/src/main/java/graphql/kickstart/servlet/CustomGraphQLServlet.java
@@ -23,8 +23,8 @@ import io.stargate.db.Persistence;
 import io.stargate.db.datastore.DataStore;
 import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.datastore.Row;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Keyspace;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Keyspace;
 import io.stargate.graphql.core.GqlKeyspaceSchema;
 import io.stargate.graphql.graphqlservlet.CassandraUnboxingGraphqlErrorHandler;
 import io.stargate.graphql.graphqlservlet.GraphqlCustomContextBuilder;
