+++ graphqlapi/src/main/java/io/stargate/graphql/core/DataFetchers.java
@@ -43,9 +43,9 @@ import io.stargate.db.QueryState;
 import io.stargate.db.datastore.DataStore;
 import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.datastore.Row;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.Table;
 import io.stargate.graphql.graphqlservlet.HTTPAwareContextImpl;
 import java.util.ArrayList;
 import java.util.Arrays;
