+++ graphqlapi/src/main/java/io/stargate/graphql/core/GqlKeyspaceSchema.java
@@ -38,9 +38,9 @@ import graphql.schema.GraphQLTypeReference;
 import graphql.schema.idl.SchemaPrinter;
 import io.stargate.auth.AuthenticationService;
 import io.stargate.db.Persistence;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.Table;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
