+++ graphqlapi/src/main/java/io/stargate/graphql/fetchers/KeyspaceFetcher.java
@@ -25,9 +25,9 @@ import io.stargate.db.ClientState;
 import io.stargate.db.Persistence;
 import io.stargate.db.QueryState;
 import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.Table;
 import io.stargate.graphql.graphqlservlet.HTTPAwareContextImpl;
 import java.util.ArrayList;
 import java.util.List;
