+++ persistence-api/src/main/java/io/stargate/db/datastore/Row.java
@@ -19,8 +19,8 @@ import com.datastax.oss.driver.api.core.data.CqlDuration;
 import com.datastax.oss.driver.api.core.data.TupleValue;
 import com.datastax.oss.driver.api.core.data.UdtValue;
 import com.datastax.oss.driver.shaded.guava.common.base.Preconditions;
-import io.stargate.db.datastore.schema.AbstractTable;
-import io.stargate.db.datastore.schema.Column;
+import io.stargate.db.schema.AbstractTable;
+import io.stargate.db.schema.Column;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.net.InetAddress;
