+++ persistence-api/src/main/java/io/stargate/db/datastore/query/QueryBuilderImpl.java
@@ -28,18 +28,18 @@ import com.github.misberner.duzzt.annotations.GenerateEmbeddedDSL;
 import com.github.misberner.duzzt.annotations.SubExpr;
 import io.stargate.db.datastore.DataStore;
 import io.stargate.db.datastore.ResultSet;
-import io.stargate.db.datastore.schema.AbstractTable;
-import io.stargate.db.datastore.schema.CollectionIndexingType;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableColumn;
-import io.stargate.db.datastore.schema.Index;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.MaterializedView;
-import io.stargate.db.datastore.schema.Schema;
-import io.stargate.db.datastore.schema.SchemaEntity;
-import io.stargate.db.datastore.schema.SecondaryIndex;
-import io.stargate.db.datastore.schema.Table;
-import io.stargate.db.datastore.schema.UserDefinedType;
+import io.stargate.db.schema.AbstractTable;
+import io.stargate.db.schema.CollectionIndexingType;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableColumn;
+import io.stargate.db.schema.Index;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.MaterializedView;
+import io.stargate.db.schema.Schema;
+import io.stargate.db.schema.SchemaEntity;
+import io.stargate.db.schema.SecondaryIndex;
+import io.stargate.db.schema.Table;
+import io.stargate.db.schema.UserDefinedType;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
