+++ persistence-api/src/main/java/io/stargate/db/schema/MaterializedView.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.db.datastore.schema;
+package io.stargate.db.schema;
 
 import com.datastax.oss.driver.shaded.guava.common.base.Preconditions;
 import java.util.List;
