+++ persistence-api/src/main/java/io/stargate/db/schema/SecondaryIndex.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.db.datastore.schema;
+package io.stargate.db.schema;
 
 import com.datastax.oss.driver.shaded.guava.common.collect.ImmutableSet;
 import io.stargate.db.datastore.query.ColumnOrder;
