+++ persistence-api/src/main/java/io/stargate/db/schema/UserDefinedType.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.stargate.db.datastore.schema;
+package io.stargate.db.schema;
 
 import static java.util.stream.Collectors.toList;
