+++ persistence-cassandra-3.11/src/main/java/io/stargate/db/cassandra/datastore/DataStoreUtil.java
@@ -23,11 +23,11 @@ import com.datastax.oss.driver.shaded.guava.common.base.Preconditions;
 import com.datastax.oss.protocol.internal.ProtocolConstants;
 import com.google.common.collect.ImmutableMap;
 import io.stargate.db.datastore.common.util.ColumnUtils;
-import io.stargate.db.datastore.schema.CollectionIndexingType;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableCollectionIndexingType;
-import io.stargate.db.datastore.schema.ImmutableColumn;
-import io.stargate.db.datastore.schema.ImmutableUserDefinedType;
+import io.stargate.db.schema.CollectionIndexingType;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableCollectionIndexingType;
+import io.stargate.db.schema.ImmutableColumn;
+import io.stargate.db.schema.ImmutableUserDefinedType;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
