+++ persistence-cassandra-3.11/src/main/java/io/stargate/db/cassandra/datastore/InternalDataStore.java
@@ -31,18 +31,18 @@ import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.datastore.common.util.ColumnUtils;
 import io.stargate.db.datastore.common.util.SchemaTool;
 import io.stargate.db.datastore.query.Parameter;
-import io.stargate.db.datastore.schema.AbstractTable;
-import io.stargate.db.datastore.schema.CollectionIndexingType;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableColumn;
-import io.stargate.db.datastore.schema.ImmutableUserDefinedType;
-import io.stargate.db.datastore.schema.Index;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.MaterializedView;
-import io.stargate.db.datastore.schema.Schema;
-import io.stargate.db.datastore.schema.SecondaryIndex;
-import io.stargate.db.datastore.schema.Table;
-import io.stargate.db.datastore.schema.UserDefinedType;
+import io.stargate.db.schema.AbstractTable;
+import io.stargate.db.schema.CollectionIndexingType;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableColumn;
+import io.stargate.db.schema.ImmutableUserDefinedType;
+import io.stargate.db.schema.Index;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.MaterializedView;
+import io.stargate.db.schema.Schema;
+import io.stargate.db.schema.SecondaryIndex;
+import io.stargate.db.schema.Table;
+import io.stargate.db.schema.UserDefinedType;
 import java.lang.reflect.Field;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
