+++ persistence-cassandra-3.11/src/main/java/io/stargate/db/cassandra/impl/Conversion.java
@@ -23,8 +23,8 @@ import io.stargate.db.Result;
 import io.stargate.db.Result.Flag;
 import io.stargate.db.cassandra.datastore.DataStoreUtil;
 import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableColumn;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableColumn;
 import java.lang.reflect.Field;
 import java.net.InetAddress;
 import java.util.ArrayList;
