+++ persistence-common/src/main/java/io/stargate/db/datastore/common/InternalRow.java
@@ -20,8 +20,8 @@ import com.datastax.oss.driver.api.core.data.TupleValue;
 import com.datastax.oss.driver.api.core.data.UdtValue;
 import io.stargate.db.datastore.Row;
 import io.stargate.db.datastore.common.util.ColumnUtils;
-import io.stargate.db.datastore.schema.AbstractTable;
-import io.stargate.db.datastore.schema.Column;
+import io.stargate.db.schema.AbstractTable;
+import io.stargate.db.schema.Column;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.net.InetAddress;
