+++ restapi/src/main/java/io/stargate/web/resources/Converters.java
@@ -28,11 +28,11 @@ import io.stargate.db.datastore.query.ImmutableValue;
 import io.stargate.db.datastore.query.ImmutableWhereCondition;
 import io.stargate.db.datastore.query.Value;
 import io.stargate.db.datastore.query.WhereCondition;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ParameterizedType;
-import io.stargate.db.datastore.schema.ReservedKeywords;
-import io.stargate.db.datastore.schema.Table;
-import io.stargate.db.datastore.schema.UserDefinedType;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ParameterizedType;
+import io.stargate.db.schema.ReservedKeywords;
+import io.stargate.db.schema.Table;
+import io.stargate.db.schema.UserDefinedType;
 import io.stargate.web.models.ClusteringExpression;
 import io.stargate.web.models.TableAdd;
 import io.stargate.web.models.TableOptions;
