+++ restapi/src/main/java/io/stargate/web/resources/Db.java
@@ -24,8 +24,8 @@ import io.stargate.db.Persistence;
 import io.stargate.db.QueryOptions;
 import io.stargate.db.QueryState;
 import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.Table;
 import java.nio.ByteBuffer;
 import java.util.Collection;
 import javax.ws.rs.NotFoundException;
