+++ restapi/src/main/java/io/stargate/web/resources/RowResource.java
@@ -20,8 +20,8 @@ import io.stargate.db.datastore.DataStore;
 import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.datastore.query.Value;
 import io.stargate.db.datastore.query.Where;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Table;
 import io.stargate.web.models.Error;
 import io.stargate.web.models.Filter;
 import io.stargate.web.models.Query;
