+++ restapi/src/main/java/io/stargate/web/resources/TableResource.java
@@ -17,9 +17,9 @@ package io.stargate.web.resources;
 
 import com.codahale.metrics.annotation.Timed;
 import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Column.Kind;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Column.Kind;
+import io.stargate.db.schema.Table;
 import io.stargate.web.models.ClusteringExpression;
 import io.stargate.web.models.ColumnDefinition;
 import io.stargate.web.models.Error;
