+++ restapi/src/main/java/io/stargate/web/resources/v2/RowsResource.java
@@ -25,8 +25,8 @@ import io.stargate.db.datastore.query.ColumnOrder;
 import io.stargate.db.datastore.query.ImmutableColumnOrder;
 import io.stargate.db.datastore.query.Value;
 import io.stargate.db.datastore.query.Where;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Table;
 import io.stargate.web.models.Error;
 import io.stargate.web.models.GetResponseWrapper;
 import io.stargate.web.models.ResponseWrapper;
