+++ restapi/src/main/java/io/stargate/web/service/WhereParser.java
@@ -23,8 +23,8 @@ import com.fasterxml.jackson.databind.ObjectReader;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.stargate.db.datastore.query.ImmutableWhereCondition;
 import io.stargate.db.datastore.query.Where;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.Table;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.Table;
 import io.stargate.web.resources.Converters;
 import java.io.IOException;
 import java.util.ArrayList;
