+++ restapi/src/test/java/io/stargate/web/resources/WhereParserTest.java
@@ -21,9 +21,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import io.stargate.db.datastore.query.ImmutableWhereCondition;
 import io.stargate.db.datastore.query.Where;
 import io.stargate.db.datastore.query.WhereCondition;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableColumn;
-import io.stargate.db.datastore.schema.ImmutableTable;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableColumn;
+import io.stargate.db.schema.ImmutableTable;
 import io.stargate.web.service.WhereParser;
 import java.io.IOException;
 import java.util.Arrays;
