+++ testing/src/test/java/io/stargate/it/PersistenceTest.java
@@ -19,35 +19,35 @@ import static com.datastax.oss.driver.api.core.uuid.Uuids.random;
 import static com.datastax.oss.driver.api.core.uuid.Uuids.timeBased;
 import static io.stargate.db.datastore.query.WhereCondition.Predicate.Eq;
 import static io.stargate.db.datastore.query.WhereCondition.Predicate.In;
-import static io.stargate.db.datastore.schema.Column.Kind.Clustering;
-import static io.stargate.db.datastore.schema.Column.Kind.PartitionKey;
-import static io.stargate.db.datastore.schema.Column.Kind.Static;
-import static io.stargate.db.datastore.schema.Column.Order.Asc;
-import static io.stargate.db.datastore.schema.Column.Order.Desc;
-import static io.stargate.db.datastore.schema.Column.Type.Ascii;
-import static io.stargate.db.datastore.schema.Column.Type.Bigint;
-import static io.stargate.db.datastore.schema.Column.Type.Blob;
-import static io.stargate.db.datastore.schema.Column.Type.Boolean;
-import static io.stargate.db.datastore.schema.Column.Type.Date;
-import static io.stargate.db.datastore.schema.Column.Type.Decimal;
-import static io.stargate.db.datastore.schema.Column.Type.Double;
-import static io.stargate.db.datastore.schema.Column.Type.Duration;
-import static io.stargate.db.datastore.schema.Column.Type.Float;
-import static io.stargate.db.datastore.schema.Column.Type.Inet;
-import static io.stargate.db.datastore.schema.Column.Type.Int;
-import static io.stargate.db.datastore.schema.Column.Type.List;
-import static io.stargate.db.datastore.schema.Column.Type.Map;
-import static io.stargate.db.datastore.schema.Column.Type.Set;
-import static io.stargate.db.datastore.schema.Column.Type.Smallint;
-import static io.stargate.db.datastore.schema.Column.Type.Text;
-import static io.stargate.db.datastore.schema.Column.Type.Time;
-import static io.stargate.db.datastore.schema.Column.Type.Timestamp;
-import static io.stargate.db.datastore.schema.Column.Type.Timeuuid;
-import static io.stargate.db.datastore.schema.Column.Type.Tinyint;
-import static io.stargate.db.datastore.schema.Column.Type.Tuple;
-import static io.stargate.db.datastore.schema.Column.Type.Uuid;
-import static io.stargate.db.datastore.schema.Column.Type.Varchar;
-import static io.stargate.db.datastore.schema.Column.Type.Varint;
+import static io.stargate.db.schema.Column.Kind.Clustering;
+import static io.stargate.db.schema.Column.Kind.PartitionKey;
+import static io.stargate.db.schema.Column.Kind.Static;
+import static io.stargate.db.schema.Column.Order.Asc;
+import static io.stargate.db.schema.Column.Order.Desc;
+import static io.stargate.db.schema.Column.Type.Ascii;
+import static io.stargate.db.schema.Column.Type.Bigint;
+import static io.stargate.db.schema.Column.Type.Blob;
+import static io.stargate.db.schema.Column.Type.Boolean;
+import static io.stargate.db.schema.Column.Type.Date;
+import static io.stargate.db.schema.Column.Type.Decimal;
+import static io.stargate.db.schema.Column.Type.Double;
+import static io.stargate.db.schema.Column.Type.Duration;
+import static io.stargate.db.schema.Column.Type.Float;
+import static io.stargate.db.schema.Column.Type.Inet;
+import static io.stargate.db.schema.Column.Type.Int;
+import static io.stargate.db.schema.Column.Type.List;
+import static io.stargate.db.schema.Column.Type.Map;
+import static io.stargate.db.schema.Column.Type.Set;
+import static io.stargate.db.schema.Column.Type.Smallint;
+import static io.stargate.db.schema.Column.Type.Text;
+import static io.stargate.db.schema.Column.Type.Time;
+import static io.stargate.db.schema.Column.Type.Timestamp;
+import static io.stargate.db.schema.Column.Type.Timeuuid;
+import static io.stargate.db.schema.Column.Type.Tinyint;
+import static io.stargate.db.schema.Column.Type.Tuple;
+import static io.stargate.db.schema.Column.Type.Uuid;
+import static io.stargate.db.schema.Column.Type.Varchar;
+import static io.stargate.db.schema.Column.Type.Varint;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Fail.fail;
 
@@ -65,14 +65,14 @@ import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.datastore.Row;
 import io.stargate.db.datastore.query.Parameter;
 import io.stargate.db.datastore.query.Value;
-import io.stargate.db.datastore.schema.Column;
-import io.stargate.db.datastore.schema.ImmutableTupleType;
-import io.stargate.db.datastore.schema.ImmutableUserDefinedType;
-import io.stargate.db.datastore.schema.Keyspace;
-import io.stargate.db.datastore.schema.ParameterizedType;
-import io.stargate.db.datastore.schema.Schema;
-import io.stargate.db.datastore.schema.Table;
-import io.stargate.db.datastore.schema.UserDefinedType;
+import io.stargate.db.schema.Column;
+import io.stargate.db.schema.ImmutableTupleType;
+import io.stargate.db.schema.ImmutableUserDefinedType;
+import io.stargate.db.schema.Keyspace;
+import io.stargate.db.schema.ParameterizedType;
+import io.stargate.db.schema.Schema;
+import io.stargate.db.schema.Table;
+import io.stargate.db.schema.UserDefinedType;
 import java.lang.reflect.Method;
 import java.math.BigDecimal;
 import java.math.BigInteger;
