+++ testing/src/test/java/io/stargate/it/http/GraphqlTest.java
@@ -43,8 +43,8 @@ import io.stargate.db.ClientState;
 import io.stargate.db.Persistence;
 import io.stargate.db.QueryState;
 import io.stargate.db.datastore.DataStore;
-import io.stargate.db.datastore.schema.Column.Kind;
-import io.stargate.db.datastore.schema.Column.Type;
+import io.stargate.db.schema.Column.Kind;
+import io.stargate.db.schema.Column.Type;
 import io.stargate.it.BaseOsgiIntegrationTest;
 import io.stargate.it.http.models.Credentials;
 import java.io.IOException;
